Extract factory into a submodule, re-export types
